backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.49
psycopg[binary]==3.2.13
alembic==1.14.0
pydantic==2.13.1
pydantic-settings==2.13.1
python-multipart==0.0.17
celery[redis]==5.4.0
redis==5.2.1
cryptography==44.0.0
stripe==11.3.0
httpx==0.28.1
boto3==1.35.76
python-dotenv==1.0.1
pillow==12.2.0
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
